Simplify plugins react router
