+++ packages/strapi-plugin-content-manager/public/app/routes.json
@@ -1,0 +1,14 @@
+{
+  "/": {
+    "name": "home",
+    "container": "Home"
+  },
+  "/:slug": {
+    "name": "list",
+    "container": "List"
+  },
+  "/:slug/:id": {
+    "name": "edit",
+    "container": "Edit"
+  }
+}
